Refactor imports to consolidate 'validator' usage across modules for improved consistency
