+++ libs/commands/registry/member.py
@@ -6,8 +6,9 @@ import logging
 from typing import TYPE_CHECKING, TypedDict, cast
 
 import libs.global_value as g
-from libs.domain import modify, validator
+from libs.domain import modify
 from libs.domain.section import BaseSection
+from libs.functions import validator
 from libs.types import CommandType
 from libs.utils import dbutil, textutil
