+++ tests/database/test_result_update.py
@@ -10,8 +10,9 @@ import pytest
 import libs.global_value as g
 from integrations import factory
 from libs.bootstrap import configuration, initialization
-from libs.domain import modify, validator
+from libs.domain import modify
 from libs.domain.score import GameResult
+from libs.functions import validator
 from libs.types import ServiceType
 from libs.utils import dbutil
 from libs.utils.timekit import ExtendedDatetime as ExtDt
